requirements.txt
mysql-connector-python==8.2.0
PyMySQL==1.1.0
telebot==0.0.5


